+++ packages/terra-tabs/src/common-tabs/_Tabs.jsx
@@ -89,7 +89,6 @@ class Tabs extends React.Component {
     this.handleDragEnd = this.handleDragEnd.bind(this);
     this.reorder = this.reorder.bind(this);
     this.resetCache();
-    // this.visibleTabs = [];
     this.state = {
       visibleTabData: this.props.tabData,
     };
@@ -113,11 +112,13 @@ class Tabs extends React.Component {
     const prevTab = prevProps.tabData.find((tab) => tab.isSelected === true);
     const currTab = this.props.tabData.find((tab) => tab.isSelected === true);
 
+    // Allow dynamic addition of tabs.
     if (this.state.visibleTabData.length !== this.props.tabData.length) {
       // eslint-disable-next-line react/no-did-update-set-state
       this.setState({ visibleTabData: this.props.tabData });
     }
 
+    // Allow Active Styles to be applied when tab is selected.
     if (prevTab.id !== currTab.id) {
       // eslint-disable-next-line react/no-did-update-set-state
       this.setState(prevArray => {